Update rule.md (#2251)
I converted the LINQ example to use modern LINQ (I forgot how the other syntax is called but it's longer and more confusing especially for more complex queries).

I'm also not sure about 2. point.
I don't understand it and I don't know from where that information comes from.
It's true that opening and closing too many connections in quick succession are a problem (it takes time for SQL Server to release connections) but EF Core LINQ still needs to make a connection to SQL Server. Maybe it's referring to that but the code above may not strictly solve it unless the original code opened and closed connection for every single query. Then yes, the time to open and close the connection for every individual query would take a lot of time.
For a dev that hasn't worked much outside ORMs, point 2 is unclear and confusing.

Also, point 2 could refer to DbContextPool, where SQL Server connection is recycled (therefore increasing the chances of SQL Server surviving a load that would otherwise kill it) and how to make sure not to prematurely run EF Core queries. (e.g. casting `IQueryable` to `IEnumerable` or `.ToList()`, etc.)
If my assumption for point 2 is correct why legacy stuff is terrible, this makes the legacy stuff even worse.