Updated Dynamics365 Roadmap image in Do you know the CRM roadmap? rule (#2613)
* Added new image for the latest Dynamics roadmap

* Added lower resolution Dynamics 365 roadmap file

* Delete Dynamics365-Roadmap.png

No suitable due to high resolution

* Delete DynamcsRoadmap.png

Image not clear.

* Re-added DynamicsRoadmap.PNG

Added a new version of DynamicsRoadmap.PNG

* Update Dynamics Roadmap image

* Deleted DynamicsRoadmap.PNG 

Had a cursor pointer in the bottom right of the screen

* Re-added DynamcisRoadmap.PNG

Added a new image without the cursor showing in the lower right

* Deleted image1.PNG

This image is deprecated. The look of the Dynamics Roadmap site landing page does not match.