Create Rule “security-email-account/rule” (#10052)
This pull request introduces a new security rule and updates the index
to include it. The most important changes involve adding a new rule
about having a security email account and updating the index to reflect
this addition.

### New Rule Addition:

*
[`rules/security-email-account/rule.md`](diffhunk://#diff-386dbf15e3aeab3f1ac0c19e42464463f832de2dcb25d394195bc7ab496a97b3R1-R46):
Added a new rule titled "Do you have a security@ email account?" which
provides guidelines on setting up and managing a security email account
to handle vulnerability reports and build trust with ethical hackers.

### Index Update:

*
[`categories/infrastructure-and-networking/rules-to-better-security.md`](diffhunk://#diff-7bab27bd060f5ba6866372fc13d383fe3ac9dfff6041d3bd01e0be4a5af86ad3R20):
Updated the index to include the new rule `security-email-account`.

---------

Co-authored-by: Jack Pettit [SSW] <57518417+JackDevAU@users.noreply.github.com>