XS ◾ Add community Discord link again (#10116)
This pull request includes a small change to the
`rules/discord-communities/rule.md` file. The change updates the
instructions for joining Discord communities by adding a hyperlink to a
list of recommended tech Discord servers.

*
[`rules/discord-communities/rule.md`](diffhunk://#diff-26b7900284a74a8732fc13bbe36758406a15bf4480b9c555afeb6cdac1d50c9bL43-R43):
Added a hyperlink to the text "join a few servers" to direct users to a
list of recommended tech Discord servers.

I paired or mob programmed with: <!-- list names or remove if not
relevant -->
✏️ @sethdaily @wicksipedia 

<!-- 
Check out the relevant rules
- https://www.ssw.com.au/rules/rules-to-better-pull-requests
- https://www.ssw.com.au/rules/write-a-good-pull-request
- https://www.ssw.com.au/rules/over-the-shoulder-prs 
- https://www.ssw.com.au/rules/do-you-use-co-creation-patterns
-->

---------

Co-authored-by: Tiago Araújo [SSW] <tiagov8@gmail.com>