Fixing expiring-app-secrets-certificates script (#10259)
I had to change our existing script from using the deprecated Azure AD
module to Graph.
I tried using the script in this rule... but it looks like it was
written by AI; some of the commands don't exist :/.

So I have written a working script and updated the rule.