XS ◾ 📈 Project portals - Updates 'as of' date for GitHub projects (#10571)
> 1. What triggered this change? (PBI link, Email Subject, conversation
+ reason, etc)

During a review of the SugarLearning Dashboard, the question was raised
as to whether GitHub projects still didn't support a Project Portal page
like that which exists in Azure DevOps.

> 2. What was changed?

After confirming that this is still the case, I've updated the 'as of'
date for the missing feature to the current date to help keep the
article up-to-date.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>